move poly into its own package
