+++ internal/poly/poly_test.go
@@ -1,4 +1,4 @@
-package kzgmulti
+package poly
 
 import (
 	"testing"
@@ -29,7 +29,7 @@ func TestPolyMul(t *testing.T) {
 func TestPolyInterpolate(t *testing.T) {
 	points := []fr.Element{fr.NewElement(1), fr.NewElement(2), fr.NewElement(3), fr.NewElement(4)}
 	values := []fr.Element{fr.NewElement(1), fr.NewElement(2), fr.NewElement(3), fr.NewElement(4)}
-	poly := Interpolate(points, values)
+	poly := LagrangeInterpolate(points, values)
 	for i, point := range points {
 		eval := PolyEval(poly, point)
 		if !eval.Equal(&values[i]) {
